Fully functional program with JUnit tests
When date entered is a valid date. This means that the entered date is within the max number of dates in the month, the entered year is between 2000 and 2025. The number of days to be added must be anywhere between the numbers 1 through 60. It also accounts for leap years.
The test cases verify the output of the program as intended, as well as the essential core functions.

this program creates an array with a size of the number of days to be added. Each element in the array includes the respectively incremented day from the entered date. It then returns the requested date which is the last element of the array.
Ex) entered 12/01/2000 want to get the date 5 days in the future:
The array then includes [12/02/2000, 12/03/2000, 12/04/2000, 12/05/2000, 12/06/2000]
The program will then return 12/6/2000.

Ex) entered 02/27/2020 want to get the date 2 days in the future:
The array then includes [02/28/2020, 02/29/2020]
The program will then return 12/29/2020.